A configurable GeneratorResolver, with DSL-based configuration
The idea is to provide an easy way to add "business"-specific conversions
by configuration, as well as to allow some variations in the code generation.

[technical]
A new package "reflect" has been created to host all those Reflection-
based classes.
Internal generators based on dynamic method resolution provide a way to
know which are the types they actually can generate code for.