Feature: Define random generation mechanisms
One often needs to generate values, test datasets for example, in a pseudo
random manner and sometimes with defined ratio applied.
For example, one might want to generate Person objects, using a dictionary
of existing names, and with a preference of 60% for women.

This feature provides basic components to achieve this a bit more easily.

Tests: Introduction of Spock Framework - works nicely for Groovy testing.